+++ CloudySkies/LayerData/ColorLayerData.cs
@@ -1,9 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 using Leclair.Stardew.Common.Serialization.Converters;
 
 using Microsoft.Xna.Framework;