+++ CloudySkies/LayerData/TextureScrollLayerData.cs
@@ -1,9 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 using Leclair.Stardew.Common.Serialization.Converters;
 
 using Microsoft.Xna.Framework;
@@ -12,6 +6,8 @@ using Newtonsoft.Json;
 
 namespace Leclair.Stardew.CloudySkies.LayerData;
 
+
+[DiscriminatedType("Snow")]
 [DiscriminatedType("TextureScroll")]
 public record TextureScrollLayerData : BaseLayerData {
 
@@ -19,13 +15,20 @@ public record TextureScrollLayerData : BaseLayerData {
 
 	public Rectangle? Source { get; set; }
 
+	public int? Frames { get; set; } // = 5;
+
+	public int TimePerFrame { get; set; } = 75;
+
 	public float Scale { get; set; } = 4;
 
 	public bool FlipHorizontal { get; set; }
 
 	public bool FlipVertical { get; set; }
 
-	public Vector2 Speed { get; set; } = Vector2.Zero;
+	public Vector2 Speed { get; set; }
+
+	public Vector2 ViewSpeed { get; set; } = new(-1, -1);
+
 
 	[JsonConverter(typeof(ColorConverter))]
 	public Color? Color { get; set; }