+++ CloudySkies/Models/CustomDebris.cs
@@ -1,5 +1,3 @@
-using System;
-
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
@@ -145,7 +143,7 @@ public class CustomDebris {
 			Vector2.Zero,
 			scale,
 			effects,
-			1E-06f
+			1E-06F
 		);
 	}