+++ Common/AdvancedMultipleMutexRequest.cs
@@ -2,13 +2,15 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-using StardewValley.Network;
+using Leclair.Stardew.Common.Events;
+
+using Netcode;
 
 using StardewModdingAPI;
-using StardewValley;
 using StardewModdingAPI.Events;
-using Leclair.Stardew.Common.Events;
-using Netcode;
+
+using StardewValley;
+using StardewValley.Network;
 
 namespace Leclair.Stardew.Common;
 
@@ -64,7 +66,7 @@ public class AdvancedMultipleMutexRequest {
 			return false;
 
 		// Double check with every mutex to ensure we hold the lock.
-		foreach(var mutex in Mutexes) {
+		foreach (var mutex in Mutexes) {
 			if (!mutex.IsLockHeld())
 				return false;
 		}
@@ -85,7 +87,7 @@ public class AdvancedMultipleMutexRequest {
 			return;
 		}
 
-		foreach(var mutex in Mutexes) {
+		foreach (var mutex in Mutexes) {
 			if (mutex.IsLocked()) {
 				OnFailure?.Invoke();
 				return;
@@ -131,7 +133,7 @@ public class AdvancedMultipleMutexRequest {
 				mutex.Update(farmers);
 
 				// See if we're held now, but didn't get an update.
-				if ( mutex.IsLockHeld() && ! AcquiredLocks.Contains(mutex)) {
+				if (mutex.IsLockHeld() && !AcquiredLocks.Contains(mutex)) {
 #if DEBUG
 					LogLevel level = LogLevel.Debug;
 #else
@@ -161,7 +163,7 @@ public class AdvancedMultipleMutexRequest {
 
 		ReportedLocks.Add(mutex);
 
-		if (! AcquiredLocks.Contains(mutex))
+		if (!AcquiredLocks.Contains(mutex))
 			AcquiredLocks.Add(mutex);
 
 		if (ReportedLocks.Count >= Mutexes.Length)
@@ -240,7 +242,7 @@ public class AdvancedMultipleMutexRequest {
 
 			Mod.LogTable(headers, states, level);
 
-		} catch(Exception) {
+		} catch (Exception) {
 			/* do nothing */
 		}
 	}
@@ -265,7 +267,7 @@ public class AdvancedMultipleMutexRequest {
 					Mod.Log($"Unable to acquire all mutexes within {Timeout} ms. IsHost: {Game1.IsMasterGame}; Multiplayer: {Context.IsMultiplayer}; Mutex state:", level);
 					LogMutexes();
 
-				} catch(Exception) {
+				} catch (Exception) {
 					/* do nothing */
 				}
 			}