+++ Common/Types/BitHelper.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Leclair.Stardew.Common.Types;